Removed cache related classes, added fat jar of cache. Removed test api classes. Updated table names, dockerized task tracker. Added GlobalHelper.java for generating token and other tasks, removed login dto. Removed redis config and dependencies. Removed 2 factor auth
