+++ build.gradle
@@ -8,7 +8,6 @@ subprojects {
     group = 'com.ab.tasktracker'
     version = '1.0'
 
-
     sourceCompatibility = 21
     targetCompatibility = 21
 
@@ -19,6 +18,11 @@ subprojects {
             extendsFrom annotationProcessor
         }
     }
+
+    tasks.withType(JavaCompile).configureEach {
+        options.compilerArgs << "-parameters"  // To preserve args name during compilation which helps in SPEL
+    }
+
 
     repositories {
         mavenCentral()
@@ -39,6 +43,5 @@ subprojects {
         testImplementation group: 'junit', name: 'junit', version: '4.12'
         testImplementation 'org.mockito:mockito-core:3.12.4'
         implementation 'com.googlecode.json-simple:json-simple:1.1.1'
-        implementation 'org.springframework.boot:spring-boot-starter-security:' + springSecurityVersion
     }
 }
