+++ gradle.properties
@@ -7,4 +7,3 @@ openFeignVersion=4.1.0
 springDocSwaggerVersion=2.5.0
 springStarterMail=3.2.3
 springCloudVersion=2023.0.0
-redisClientJedis=5.1.3
