+++ resources/application.properties
@@ -7,26 +7,25 @@ app.name=TaskTracker
 app.description=Application Task Tracker
 spring.application.name=TaskTracker
 spring.application.description=Application Task Tracker
-spring.main.allow-circular-references=true
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
 server.servlet.register-default-servlet=true
 logging.level.com.ab.tasktracker=debug
-server.address=https://thespecialcharacter.in/
+server.host=localhost
 server.port=8090
 current.database=pgsql
 ###########################################################################
 #PostgreSQL Database
 ###########################################################################
 spring.datasource.pgsql.url=jdbc:postgresql://localhost:5432/ABTaskTrackerDB
-spring.datasource.pgsql.username=postgres
-spring.datasource.pgsql.password=root
+spring.datasource.pgsql.username=${POSTGRES_USERNAME}
+spring.datasource.pgsql.password=${POSTGRES_PASSWORD}
 spring.jpa.properties.pgsql.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 ###########################################################################
 #MYSQL Database
 ###########################################################################
 spring.datasource.mysql.url=jdbc:mysql://localhost:3306/ABTaskTrackerDB
-spring.datasource.mysql.username=root
-spring.datasource.mysql.password=root
+spring.datasource.mysql.username=${MYSQL_USERNAME}
+spring.datasource.mysql.password=${MYSQL_PASSWORD}
 spring.jpa.properties.mysql.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 ###########################################################################
 #JPA Properties
@@ -56,9 +55,10 @@ spring.datasource.hikari.auto-commit=true
 #############################################################################
 #JWT
 ###########################################################################
-jwt.secret.key=6qr6%hty&t)j8xxphdl$go+kn1v4+8b89c6_mn_sq7%v1yrujhj8xxphdl$go+kn1v4+8b89c6_mn_sq7%v1yrujh
+jwt.secret.key=${JWT_SECRET_KEY}
 jwt.expiration.time=86400000
 logging.level.org.springframework.security=TRACE
+services.list=auth-service,notification
 #############################################################################
 #XSS
 ###########################################################################
@@ -67,9 +67,8 @@ xss.invalid.literals=javas&#99;ript,javascript:,<script,</script>,vbscript:,live
 #Redis
 ###########################################################################
 spring.cache.type=redis
-spring.cache.host=localhost
-spring.cache.port=6380
-redis.nodes=127.0.0.1:6379,127.0.0.1:6380,127.0.0.1:6381
+redis.nodes=127.0.0.1:6378,127.0.0.1:6380,127.0.0.1:6381
+#1200 milliseconds = 20 min
 spring.cache.redis.time-to-live=1200000
 logging.level.org.springframework.data.redis=DEBUG
 logging.level.org.springframework.data.redis.core=DEBUG
