+++ resources/redis.txt
@@ -15,11 +15,14 @@ https://stackoverflow.com/questions/75182048/redis-cluster-setup-on-ubuntu-20-04
 
 1.3)Make cluster -
 
-redis-cli --cluster create 127.0.0.1:6379 127.0.0.1:6380 127.0.0.1:6381 --cluster-replicas 0
+redis-cli --cluster create 127.0.0.1:6378 127.0.0.1:6380 127.0.0.1:6381 --cluster-replicas 0
 
 1.4)sudo service redis-server stop
 
 1.5)Reset clusters
+
+1.6)Check cluster nodes
+redis-cli -p 6380 cluster nodes
 
 redis-cli -p <port>
