--- tasktracker-client/src/main/java/com/ab/tasktracker/client/TwoFactorAuthClient.java
@@ -1,7 +1,0 @@
-package com.ab.tasktracker.client;
-
-import org.springframework.cloud.openfeign.FeignClient;
-
-@FeignClient(url = "localhost:8181/2fa/")
-public interface TwoFactorAuthClient {
-}
