+++ tasktracker-client/src/main/java/enums/MailType.java
@@ -1,0 +1,10 @@
+package enums;
+
+public enum MailType {
+    FORGOT_PASSWORD_MAIL("ForgotPassword"),
+    VALIDATE_SIGNUP_MAIL("ValidateSignUp"),
+    NEW_DEVICE_LOGIN_MAIL("NewDeviceLogin");
+
+    MailType(String s) {
+    }
+}
