+++ tasktracker-core/src/main/java/com/ab/tasktracker/constants/TaskTrackerURI.java
@@ -1,21 +1,12 @@
 package com.ab.tasktracker.constants;
 
 public interface TaskTrackerURI {
-    String USER_URI = "user";
     String TASK_URI = "task";
     String TEST_URI = "test";
-    String SIGN_UP_URI = "/signup";
-    String VALIDATE_USER_SIGNUP_URI = "/twofactorsignup";
-    String LOGIN_URI = "/login";
-    String API_URI = "/api";
-    String ME_URI = "/me";
     String ADD_TASK_URI = "/addtask";
     String GET_TASK_URI = "/gettask";
     String GET_ALL_ME_TASK_URI = "/getAllTasks";
     String REMOVE_TASK_URI = "/removetask";
-    String FORGOT_PASSWORD_URI = "/forgot-password";
-    String VALIDATE_OTP_URI = "/validate-otp";
-    String CHANGE_PASSWORD_URI = "/change-password";
     String TRANS_TEST_GET_API_URI = "/test-get-api";
     String CACHE_TEST_API_URI = "/test-cache-api";
     String TYPESENSE_TEST_API_URI = "/test-typesense-api";
