+++ tasktracker-dao/build.gradle
@@ -2,6 +2,4 @@
 dependencies {
     api 'org.springframework.boot:spring-boot-starter-data-jpa'
     runtimeOnly 'org.postgresql:postgresql'
-    api 'org.springframework.boot:spring-boot-starter-data-redis'
-    api 'redis.clients:jedis'
 }
