--- tasktracker-dao/src/main/java/com/ab/tasktracker/enums/MailType.java
@@ -1,10 +1,0 @@
-package com.ab.tasktracker.enums;
-
-public enum MailType {
-    FORGOT_PASSWORD_MAIL("ForgotPassword"),
-    VALIDATE_SIGNUP_MAIL("ValidateSignUp"),
-    NEW_DEVICE_LOGIN_MAIL("NewDeviceLogin");
-
-    MailType(String s) {
-    }
-}
