+++ tasktracker-typesense/build.gradle
@@ -2,4 +2,5 @@
 dependencies {
     api 'org.typesense:typesense-java:0.8.1'
     api 'com.fasterxml.jackson.core:jackson-databind:2.12.3'
+    api 'org.springframework.boot:spring-boot-starter-web'
 }
